code/requirements.txt
# ============================================================
# requirements.txt — Fake Image Detection Pipeline
# ============================================================
# Install with: pip install -r requirements.txt
# Python >= 3.9 required
# CUDA >= 11.8 recommended for GPU training

# ── Core deep learning ────────────────────────────────────
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# ── ViT / DINO (loaded via torch.hub, no extra install needed)
# torch.hub.load('facebookresearch/dino:main', 'dino_vitb16') 
# requires internet access on first run

# ── Face detection ────────────────────────────────────────
retina-face>=0.0.14       # RetinaFace detector
opencv-python>=4.8.0      # Image I/O and geometric transforms

# ── Data science ─────────────────────────────────────────
numpy>=1.24.0
scikit-learn>=1.3.0       # roc_auc_score, average_precision_score
Pillow>=10.0.0            # PIL image loading

# ── Utilities ─────────────────────────────────────────────
tqdm>=4.66.0              # progress bars
matplotlib>=3.7.0         # visualisation in predict.py

# ── Optional: faster data loading ────────────────────────
# albumentations>=1.3.0   # alternative augmentation library
